+++ mobile/lib/ui/account/change_email_dialog.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import "package:photos/l10n/l10n.dart";
-import 'package:photos/services/user_service.dart';
+import 'package:photos/services/account/user_service.dart';
 import 'package:photos/utils/dialog_util.dart';
 import 'package:photos/utils/email_util.dart';
