+++ mobile/lib/ui/account/login_pwd_verification_page.dart
@@ -6,7 +6,7 @@ import "package:logging/logging.dart";
 import 'package:photos/core/configuration.dart';
 import "package:photos/generated/l10n.dart";
 import "package:photos/models/api/user/srp.dart";
-import "package:photos/services/user_service.dart";
+import "package:photos/services/account/user_service.dart";
 import "package:photos/theme/ente_theme.dart";
 import 'package:photos/ui/common/dynamic_fab.dart';
 import "package:photos/ui/components/buttons/button_widget.dart";
