+++ mobile/lib/ui/account/password_reentry_page.dart
@@ -9,7 +9,7 @@ import 'package:photos/core/errors.dart';
 import 'package:photos/core/event_bus.dart';
 import 'package:photos/events/subscription_purchased_event.dart';
 import "package:photos/generated/l10n.dart";
-import "package:photos/services/user_service.dart";
+import "package:photos/services/account/user_service.dart";
 import "package:photos/theme/ente_theme.dart";
 import 'package:photos/ui/account/recovery_page.dart';
 import 'package:photos/ui/common/dynamic_fab.dart';
