+++ mobile/lib/ui/account/two_factor_recovery_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import "package:photos/generated/l10n.dart";
 import "package:photos/models/account/two_factor.dart";
-import 'package:photos/services/user_service.dart';
+import 'package:photos/services/account/user_service.dart';
 import "package:photos/theme/ente_theme.dart";
 import 'package:photos/utils/dialog_util.dart';
