+++ mobile/lib/ui/account/verify_recovery_page.dart
@@ -8,8 +8,8 @@ import 'package:photos/ente_theme_data.dart';
 import 'package:photos/events/notification_event.dart';
 import "package:photos/generated/l10n.dart";
 import "package:photos/service_locator.dart";
+import 'package:photos/services/account/user_service.dart';
 import 'package:photos/services/local_authentication_service.dart';
-import 'package:photos/services/user_service.dart';
 import "package:photos/theme/ente_theme.dart";
 import 'package:photos/ui/account/recovery_key_page.dart';
 import 'package:photos/ui/common/gradient_button.dart';
