+++ mobile/lib/ui/payment/payment_web_page.dart
@@ -10,8 +10,8 @@ import 'package:photos/ente_theme_data.dart';
 import "package:photos/generated/l10n.dart";
 import 'package:photos/models/api/billing/subscription.dart';
 import "package:photos/service_locator.dart";
-import 'package:photos/services/billing_service.dart';
-import 'package:photos/services/user_service.dart';
+import 'package:photos/services/account/billing_service.dart';
+import 'package:photos/services/account/user_service.dart';
 import 'package:photos/ui/common/loading_widget.dart';
 import 'package:photos/utils/dialog_util.dart';
 import "package:photos/utils/email_util.dart";
