+++ mobile/lib/ui/payment/store_subscription_page.dart
@@ -13,7 +13,7 @@ import 'package:photos/models/api/billing/billing_plan.dart';
 import 'package:photos/models/api/billing/subscription.dart';
 import 'package:photos/models/user_details.dart';
 import "package:photos/service_locator.dart";
-import 'package:photos/services/user_service.dart';
+import 'package:photos/services/account/user_service.dart';
 import "package:photos/theme/colors.dart";
 import "package:photos/theme/ente_theme.dart";
 import 'package:photos/ui/common/loading_widget.dart';
