+++ mobile/lib/ui/settings/account_section_widget.dart
@@ -5,8 +5,8 @@ import "package:flutter/foundation.dart";
 import 'package:flutter/material.dart';
 import "package:photos/emergency/emergency_page.dart";
 import "package:photos/generated/l10n.dart";
+import 'package:photos/services/account/user_service.dart';
 import 'package:photos/services/local_authentication_service.dart';
-import 'package:photos/services/user_service.dart';
 import 'package:photos/theme/ente_theme.dart';
 import 'package:photos/ui/account/change_email_dialog.dart';
 import 'package:photos/ui/account/delete_account_page.dart';
