+++ mobile/lib/ui/settings/general_section_widget.dart
@@ -5,7 +5,7 @@ import "package:photos/app.dart";
 import "package:photos/generated/l10n.dart";
 import "package:photos/l10n/l10n.dart";
 import "package:photos/service_locator.dart";
-import 'package:photos/services/user_service.dart';
+import 'package:photos/services/account/user_service.dart';
 import 'package:photos/theme/ente_theme.dart';
 import "package:photos/ui/components/captioned_text_widget.dart";
 import "package:photos/ui/components/expandable_menu_item_widget.dart";
