+++ mobile/lib/ui/tabs/home_widget.dart
@@ -34,13 +34,13 @@ import 'package:photos/models/collection/collection_items.dart';
 import "package:photos/models/file/file.dart";
 import 'package:photos/models/selected_files.dart';
 import "package:photos/service_locator.dart";
+import 'package:photos/services/account/user_service.dart';
 import 'package:photos/services/app_lifecycle_service.dart';
 import 'package:photos/services/collections_service.dart';
 import 'package:photos/services/local_sync_service.dart';
 import "package:photos/services/machine_learning/face_ml/person/person_service.dart";
 import "package:photos/services/notification_service.dart";
 import "package:photos/services/remote_sync_service.dart";
-import 'package:photos/services/user_service.dart';
 import 'package:photos/states/user_details_state.dart';
 import 'package:photos/theme/colors.dart';
 import "package:photos/theme/effects.dart";
