+++ mobile/lib/utils/file_uploader.dart
@@ -34,11 +34,11 @@ import "package:photos/models/user_details.dart";
 import 'package:photos/module/upload/model/upload_url.dart';
 import "package:photos/module/upload/service/multipart.dart";
 import "package:photos/service_locator.dart";
+import "package:photos/services/account/user_service.dart";
 import 'package:photos/services/collections_service.dart';
 import 'package:photos/services/local_sync_service.dart';
 import "package:photos/services/preview_video_store.dart";
 import 'package:photos/services/sync_service.dart';
-import "package:photos/services/user_service.dart";
 import 'package:photos/utils/data_util.dart';
 import "package:photos/utils/exif_util.dart";
 import "package:photos/utils/file_key.dart";
